AR-1224: export render obj function
